move networkId types module
