+++ src/apps/curve/positions.e2e.ts
@@ -1,5 +1,5 @@
 import hook from './positions'
-import { NetworkId } from '../../api/networkId'
+import { NetworkId } from '../../types/networkId'
 
 describe('getPositionDefinitions', () => {
   it.each([NetworkId['ethereum-mainnet'], NetworkId['celo-mainnet']])(
