+++ src/apps/locked-celo/positions.ts
@@ -11,7 +11,7 @@ import {
 import { celo } from 'viem/chains'
 import { LockedGoldAbi } from './abis/locked-gold'
 import { toDecimalNumber } from '../../types/numbers'
-import { NetworkId } from '../../api/networkId'
+import { NetworkId } from '../../types/networkId'
 
 const CELO_ADDRESS = '0x471ece3750da237f93b8e339c536989b8978a438'
 const LOCKED_GOLD_ADDRESS = '0x6cc083aed9e3ebe302a6336dbc7c921c9f03349e'
