+++ src/runtime/isNative.test.ts
@@ -1,4 +1,4 @@
-import { NetworkId } from '../api/networkId'
+import { NetworkId } from '../types/networkId'
 import { isNative } from './isNative'
 
 describe('isNative', () => {
